requirements/main.txt
Django==2.2.14
envparse==0.2.0
gunicorn==20.0.4
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
redis==3.4.1
celery==4.4.7
django-celery-results==1.2.1
requests==2.23.0
urllib3==1.25.8
djangorestframework==3.11.0
sentry-sdk==0.19.1
beautifulsoup4==4.9.0
django-prometheus==2.1.0
fluent-logger==0.9.6
html5lib==1.0.1
CacheControl==0.12.6